+++ MercadoPagoSDK/MercadoPagoSDK/Config/PXPaymentConfiguration.swift
@@ -35,8 +35,9 @@ extension PXPaymentConfiguration {
      Add your own payment method option to pay.
      - parameter plugin: Your custom payment method plugin.
      */
+    @available(*, deprecated: 4.5.0, message: "Payment method plugins is no longer available.")
+    /// :nodoc
     open func addPaymentMethodPlugin(plugin: PXPaymentMethodPlugin) -> PXPaymentConfiguration {
-        self.paymentMethodPlugins.append(plugin)
         return self
     }