+++ MercadoPagoSDK/MercadoPagoSDK/Core/MercadoPagoServicesAdapter.swift
@@ -44,7 +44,7 @@ internal class MercadoPagoServicesAdapter {
 
     typealias PaymentSearchExclusions = (excludedPaymentTypesIds: [String], excludedPaymentMethodsIds: [String])
     typealias PaymentSearchOneTapInfo = (cardsWithEsc: [String]?, supportedPlugins: [String]?)
-    typealias ExtraParams = (defaultPaymentMethod: String?, differentialPricingId: String?, defaultInstallments: String?, expressEnabled: Bool)
+    typealias ExtraParams = (defaultPaymentMethod: String?, differentialPricingId: String?, defaultInstallments: String?, expressEnabled: Bool, hasPaymentProcessor: Bool)
 
     func getPaymentMethodSearch(amount: Double, exclusions: PaymentSearchExclusions, oneTapInfo: PaymentSearchOneTapInfo, payer: PXPayer, site: String, extraParams: ExtraParams?, discountParamsConfiguration: PXDiscountParamsConfiguration?, marketplace: String?, charges: [PXPaymentTypeChargeRule]?, callback : @escaping (PXPaymentMethodSearch) -> Void, failure: @escaping ((_ error: NSError) -> Void)) {
 
@@ -52,16 +52,16 @@ internal class MercadoPagoServicesAdapter {
 
         let pxSite = getPXSiteFromId(site)
 
-        let accountMoneyAvailable: Bool = false // TODO: This is temporary. (Warning: Until AM First Class Member)
-
-        var excludedPaymentTypesIds = exclusions.excludedPaymentTypesIds
-        if !accountMoneyAvailable {
-            excludedPaymentTypesIds.append("account_money")
-        }
-
         var expressValue: String = "false"
         if let eParams = extraParams, eParams.expressEnabled {
             expressValue = "true"
+        }
+
+        var excludedPaymentTypesIds = exclusions.excludedPaymentTypesIds
+        if let eParams = extraParams, !eParams.hasPaymentProcessor {
+        // Only until our backend can pay with account money.
+        // Add exclusion for account money.
+        excludedPaymentTypesIds.append(PXPaymentTypes.ACCOUNT_MONEY.rawValue)
         }
 
         mercadoPagoServices.getPaymentMethodSearch(amount: amount, excludedPaymentTypesIds: exclusions.excludedPaymentTypesIds, excludedPaymentMethodsIds: exclusions.excludedPaymentMethodsIds, cardsWithEsc: oneTapInfo.cardsWithEsc, supportedPlugins: oneTapInfo.supportedPlugins, defaultPaymentMethod: extraParams?.defaultPaymentMethod, payer: payer, site: pxSite, differentialPricingId: extraParams?.differentialPricingId, defaultInstallments: extraParams?.defaultInstallments, expressEnabled: expressValue, discountParamsConfiguration: discountParamsConfiguration, marketplace: marketplace, charges: charges, callback: { (pxPaymentMethodSearch) in