+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXCustomOptionSearchItem.swift
@@ -17,9 +17,10 @@ open class PXCustomOptionSearchItem: NSObject, Codable {
     open var selectedAmountConfiguration: String
     open var selectedPayerCostConfiguration: PXPayerCostConfiguration?
     open var payerCostConfigurations: [String: PXPayerCostConfiguration]
+    open var comment: String?
 
 
-    public init(id: String, description: String?, paymentMethodId: String?, paymentTypeId: String?, discountInfo: String?, selectedAmountConfiguration: String, payerCostConfigurations: [String: PXPayerCostConfiguration]) {
+    public init(id: String, description: String?, paymentMethodId: String?, paymentTypeId: String?, discountInfo: String?, selectedAmountConfiguration: String, payerCostConfigurations: [String: PXPayerCostConfiguration], comment: String?) {
         self.id = id
         self._description = description
         self.paymentMethodId = paymentMethodId
@@ -27,6 +28,7 @@ open class PXCustomOptionSearchItem: NSObject, Codable {
         self.discountInfo = discountInfo
         self.selectedAmountConfiguration = selectedAmountConfiguration
         self.payerCostConfigurations = payerCostConfigurations
+        self.comment = comment
 
         if let selectedPayerCostConfiguration = payerCostConfigurations[selectedAmountConfiguration] {
             self.selectedPayerCostConfiguration = selectedPayerCostConfiguration
@@ -41,6 +43,7 @@ open class PXCustomOptionSearchItem: NSObject, Codable {
         case discountInfo = "discount_info"
         case selectedAmountConfiguration = "selected_amount_configuration"
         case payerCostConfigurations = "payer_cost_configurations"
+        case comment = "comment"
     }
 
     required public convenience init(from decoder: Decoder) throws {
@@ -49,11 +52,12 @@ open class PXCustomOptionSearchItem: NSObject, Codable {
         let description: String? = try container.decodeIfPresent(String.self, forKey: .description)
         let paymentMethodId: String? = try container.decodeIfPresent(String.self, forKey: .paymentMethodId)
         let paymentTypeId: String? = try container.decodeIfPresent(String.self, forKey: .paymentTypeId)
+        let comment: String? = try container.decodeIfPresent(String.self, forKey: .comment)
         let discountInfo: String? = try container.decodeIfPresent(String.self, forKey: .discountInfo)
         let selectedAmountConfiguration: String = try container.decode(String.self, forKey: .selectedAmountConfiguration)
         let payerCostConfigurations: [String: PXPayerCostConfiguration] = try container.decode([String: PXPayerCostConfiguration].self, forKey: .payerCostConfigurations)
 
-        self.init(id: id, description: description, paymentMethodId: paymentMethodId, paymentTypeId: paymentTypeId, discountInfo: discountInfo, selectedAmountConfiguration: selectedAmountConfiguration, payerCostConfigurations: payerCostConfigurations)
+        self.init(id: id, description: description, paymentMethodId: paymentMethodId, paymentTypeId: paymentTypeId, discountInfo: discountInfo, selectedAmountConfiguration: selectedAmountConfiguration, payerCostConfigurations: payerCostConfigurations, comment: comment)
     }
 
     public func encode(to encoder: Encoder) throws {
@@ -65,6 +69,7 @@ open class PXCustomOptionSearchItem: NSObject, Codable {
         try container.encodeIfPresent(self.discountInfo, forKey: .discountInfo)
         try container.encodeIfPresent(self.selectedAmountConfiguration, forKey: .selectedAmountConfiguration)
         try container.encodeIfPresent(self.payerCostConfigurations, forKey: .payerCostConfigurations)
+        try container.encodeIfPresent(self.comment, forKey: .comment)
     }
 
     open func toJSONString() throws -> String? {