+++ MercadoPagoSDK/MercadoPagoSDK/UI/PaymentVault/PaymentOptionDrawable.swift
@@ -34,4 +34,6 @@ internal protocol PaymentMethodOption {
     func isCard() -> Bool
 
     func isCustomerPaymentMethod() -> Bool
+
+    func getPaymentType() -> String
 }